requirements.txt
Django>=6.0,<7.0
django-allauth>=0.63
django-environ>=0.11
django-cors-headers>=4.4
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
whitenoise>=6.7
psycopg2-binary>=2.9
Pillow>=10.3
PyJWT>=2.8
cryptography>=42.0
requests>=2.32
